aeropuerto-cordoba-dashboard/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
